FISH-7957 remove trailing / from JAVA_HOME, JDK11_HOME, it fails jacc and webservices12 TCKs
The trailing / gets to the path to java. When the script changes
Glassfish port, the command stop-domain fails as it doesn't connect to
the server. The next command tries to kill all java processes, but
grepping to the path with "...java//bin/java", obviously missing it.
Then it tries to start Payara, which fails because of occupied ports.